Fixed import of js file
There's no need for curly braces since there is only one default export in the ui.js module.
Also added a rule in the eslintrc.json file that enforces the need to put file extensions to a module's path. This fixes the server not being able to serve js files due to not recognizing them.